Toggle login/logout on nav bar
