+++ src/client/app/components/nav-bar.tsx
@@ -1,10 +1,16 @@
 import React from 'react';
-import { Link, useLocation } from 'react-router-dom';
+import { Link, useHistory, useLocation } from 'react-router-dom';
+import { useRecoilValue } from 'recoil';
+import { tokenState } from 'src/client/login/state/atoms';
+import { useGoogleSignIn } from 'src/client/team/hooks/use-google-sign-in';
 import { StyledListItem, StyledNav } from './styled';
 
 export const NavBar = ({ signOut }: { signOut: () => void }) => {
   const location = useLocation();
   const currentRoute = location.pathname;
+  const token = useRecoilValue(tokenState);
+  const history = useHistory();
+  const { signIn } = useGoogleSignIn(() => history.push('team'));
 
   return (
     <StyledNav>
@@ -16,9 +22,15 @@ export const NavBar = ({ signOut }: { signOut: () => void }) => {
           <Link to='/leaderboard'>Leaderboard</Link>
         </StyledListItem>
         <StyledListItem active={currentRoute === '/login'}>
-          <Link to='/leaderboard' onClick={signOut}>
-            Logout
-          </Link>
+          {token ? (
+            <Link to='/leaderboard' onClick={signOut}>
+              Logout
+            </Link>
+          ) : (
+            <Link to='/team' onClick={signIn}>
+              Login
+            </Link>
+          )}
         </StyledListItem>
       </ul>
     </StyledNav>
